- Added support for damage-specific player pain sounds. - Removed the constraint that all $playerreserve SNDINFO commands must come   before the other $player commands. - Fixed: TArray::Reserve did not construct its newly allocated entries. - Changed the damage type for drowning from 'Water' to 'Drowning'.
SVN r466 (trunk)

+++ src/w_wad.h
@@ -133,7 +133,7 @@ public:
 	FMemLump &operator= (const FMemLump &copy);
 	~FMemLump ();
 	void *GetMem () { return Block.Len() == 0 ? NULL : (void *)Block.GetChars(); }
-	unsigned int GetSize () { return Block.Len(); }
+	size_t GetSize () { return Block.Len(); }
 
 private:
 	FMemLump (const FString &source);
